+++ test/submit.react.test.js
@@ -37,9 +37,10 @@ test('Submits', async () => {
   const mockSubmit = jest.fn(() => {
     doSubmit()
   })
+  const mockChange = jest.fn()
 
   const app = mount(
-    <Autoform schema={custom} onSubmit={mockSubmit} />
+    <Autoform schema={custom} onSubmit={mockSubmit} onChange={mockChange} />
   )
 
   const input = app.find('input[name="name"]')
@@ -57,21 +58,28 @@ test('Submits', async () => {
   const booler = app.find('input[name="booler"]')
   changeCheckbox(booler, true)
 
+  const wantedDoc = {
+    name: 'Hello',
+    numberer: 128,
+    radiator: 'b',
+    selector: 'd',
+    booler: true
+  }
+
+  // Test onChanges
+  expect(mockChange.mock.calls.length).toBe(5)
+  expect(mockChange.mock.calls[4][0]).toStrictEqual(wantedDoc)
+
+  // Test submit
   const Form = app.find(Autoform)
   await act(async () => {
     await Form.simulate('submit')
   })
 
-  expect.assertions(2)
+  expect.assertions(4)
   const { calls } = mockSubmit.mock
   return wasSubmitted.then(() => {
     expect(calls.length).toBe(1)
-    expect(calls[0][0]).toStrictEqual({
-      name: 'Hello',
-      numberer: 128,
-      radiator: 'b',
-      selector: 'd',
-      booler: true
-    })
+    expect(calls[0][0]).toStrictEqual(wantedDoc)
   })
 })